refactor: modify express routes
