+++ routes/modules/hello.js
@@ -1,0 +1,8 @@
+const express = require('express')
+const router = express.Router()
+
+router.get('/hello', (req, res) => {
+  return res.send('Hello World')
+})
+
+module.exports = router
